+++ src/models/personalized.ts
@@ -1,0 +1,13 @@
+export interface PersonalizedPlaylist {
+  alg: string
+  canDislike: boolean
+  copywriter: string
+  highQuality: boolean
+  id: number
+  name: string
+  picUrl: string
+  playCount: number
+  trackCount: number
+  trackNumberUpdateTime: number
+  type: number
+}