+++ src/router/pages.ts
@@ -3,6 +3,7 @@ export const Pages = {
   home: 'home',
   migration: 'migration',
   notFound: 'notFound',
+  playlist: 'playlist',
 } as const
 
 export type PageName = (typeof Pages)[keyof typeof Pages]